Uploading my Toolbox_Assignment_ NLP  ElAllam
Natural Language Understanding is a collection of APIs that offer text analysis through natural language processing. This set of APIs can analyze text to help you understand its concepts, entities, keywords, sentiment, and more. Additionally, you can analyse a large quantity of text in no time at all.,you can create a custom model for some APIs to get specific results that are tailored to your domain (journalism,, research, and specific text understanding), it could also be used  in more of a marketing application, you  can  also use this to find out which themes or topics users are referring to when they give their thoughts about a product or service for example to refine a commercial pitch.This type of analysis can be very useful in customer support, to analyse user requests and, respond the angriest users first.